+++ src/batou/migrate/tests/test_migrate.py
@@ -1,0 +1,116 @@
+import json
+import os
+import sys
+import textwrap
+from unittest import mock
+
+import pytest
+
+from .. import CONFIG_FILE_NAME, main, migrate, read_config, write_config
+
+
+def test_migrate__read_config__1(tmp_path):
+    """It returns the migration configuration version from ``.batou.json.``."""
+    (tmp_path / CONFIG_FILE_NAME).write_text(
+        json.dumps({'migration': {
+            'version': 2300}}))
+    os.chdir(tmp_path)
+    assert read_config() == 2300
+
+
+def test_migrate__read_config__2(tmp_path):
+    """It raises a FileNotFoundError if the configuration file is missing."""
+    with pytest.raises(FileNotFoundError):
+        read_config()
+
+
+def test_migrate__read_config__3(tmp_path):
+    """It raises a KeyError if the configuration file does not have the ...
+
+    right structure.
+    """
+    (tmp_path / CONFIG_FILE_NAME).write_text(json.dumps({'version': 2300}))
+    os.chdir(tmp_path)
+    with pytest.raises(KeyError):
+        assert read_config()
+
+
+def test_migrate__read_config__4(tmp_path):
+    """It converts the migration version number to an integer."""
+    (tmp_path / CONFIG_FILE_NAME).write_text(
+        json.dumps({'migration': {
+            'version': "2411"}}))
+    os.chdir(tmp_path)
+    assert read_config() == 2411
+
+
+@pytest.fixture(scope='function')
+def migrations(tmp_path):
+    """Create some simple migrations."""
+    TEMPLATE = textwrap.dedent("""
+        def migrate():
+            print({})
+    """)
+    os.chdir(tmp_path)
+    package = (tmp_path / 'package')
+    package.mkdir()
+    (package / '__init__.py').touch()
+    migrations = (package / 'migrations')
+    migrations.mkdir()
+    (migrations / '__init__.py').touch()
+    for step in (2411, 2301, 2300, 2299):
+        (migrations / f'{step}.py').write_text(TEMPLATE.format(step))
+    sys.path.insert(0, str(tmp_path))
+    with mock.patch('batou.migrate.MIGRATION_MODULE', new='package'):
+        yield
+    sys.path.pop()
+
+
+def test_migrate__migrate__1(migrations, capsys):
+    """It runs all migration steps starting from next after the ...
+
+    given version.
+    """
+    assert migrate(2300) == 2411
+    assert '2301\n2411\n' == capsys.readouterr().out
+
+
+def test_migrate__migrate__2(migrations, capsys):
+    """It does nothing if we are already on the latest migration step."""
+    assert migrate(2411) == 2411
+    assert '' == capsys.readouterr().out
+
+
+def test_migrate__write_config__1(tmp_path):
+    """It writes the configuration file."""
+    os.chdir(tmp_path)
+    write_config(2300)
+    assert (tmp_path / CONFIG_FILE_NAME).exists()
+    assert read_config() == 2300
+
+
+def test_migrate__write_config__2(tmp_path):
+    """It overwrites an existing the configuration file."""
+    os.chdir(tmp_path)
+    write_config(2300)
+    write_config(2311)
+    assert read_config() == 2311
+
+
+def test_migrate__main__1(tmp_path, migrations, capsys):
+    """It runs all migrations if no configuration file is present.
+
+    It creates a configuration file.
+    """
+    main()
+    assert '2299\n2300\n2301\n2411\n' == capsys.readouterr().out
+    assert (tmp_path / CONFIG_FILE_NAME).exists()
+    assert read_config() == 2411
+
+
+def test_migrate__main__2(tmp_path, migrations, capsys):
+    """It runs no migrations if already at the latest migration step."""
+    write_config(2411)
+    main()
+    assert '' == capsys.readouterr().out
+    assert read_config() == 2411